requirements.txt
SQLAlchemy>=1.3.18
aPRAW>=0.5.3a0
discord.py>=1.3.4
